requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pdfplumber
pandas
pypdf
SpeechRecognition
pyttsx3